folder restructuring & refactored helper classes into static classes - new is glue?
