+++ MarsRoverKata/MarsRoverKata.csproj
@@ -6,11 +6,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
   <ItemGroup>
-    <None Remove="Mappers\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Mappers\" />
+    <Folder Include="Helpers\" />
+    <Folder Include="Models\" />
   </ItemGroup>
 </Project>
