+++ MarsRoverKata/MarsRoverKata.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 25.0.1705.5
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarsRoverKata", "MarsRoverKata.csproj", "{16BBBFDD-BBDD-4539-B382-1AAB16AB3D24}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{16BBBFDD-BBDD-4539-B382-1AAB16AB3D24}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{16BBBFDD-BBDD-4539-B382-1AAB16AB3D24}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16BBBFDD-BBDD-4539-B382-1AAB16AB3D24}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{16BBBFDD-BBDD-4539-B382-1AAB16AB3D24}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {3B7F70B4-A39B-48EC-99D2-CABE5B977154}
+	EndGlobalSection
+EndGlobal
